+++ packages/user/src/components/ChangePasswordForm/index.tsx
@@ -3,6 +3,7 @@ import { useTranslation } from "@prefabs.tech/react-i18n";
 import React from "react";
 import * as zod from "zod";
 
+import { useConfig } from "@/hooks";
 import { ChangePasswordFormData } from "@/views/ChangePassword";
 
 import ChangePasswordFormFields from "./ChangePasswordFormFields";
@@ -18,25 +19,29 @@ interface Properties {
 
 export const ChangePasswordForm = ({ handleSubmit, loading }: Properties) => {
   const { t, i18n } = useTranslation("user");
+  const config = useConfig();
+  const hasConfirmPasswordFeature = config?.features?.confirmPassword ?? false;
 
-  const ChangePasswordFormSchema = zod
-    .object({
-      oldPassword: zod
-        .string()
-        .nonempty(t("changePassword.messages.validation.currentPassword")),
-      ...PasswordConfirmationSchema({
-        passwordValidationMessage: t(
-          "changePassword.messages.validation.mustContain",
-        ),
-        passwordRequiredMessage: t(
-          "changePassword.messages.validation.newPassword",
-        ),
-        confirmPasswordRequiredMessage: t(
-          "changePassword.messages.validation.confirmPassword",
-        ),
-      }),
-    })
-    .refine(
+  let ChangePasswordFormSchema = zod.object({
+    oldPassword: zod
+      .string()
+      .nonempty(t("changePassword.messages.validation.currentPassword")),
+    ...PasswordConfirmationSchema({
+      passwordValidationMessage: t(
+        "changePassword.messages.validation.mustContain",
+      ),
+      passwordRequiredMessage: t(
+        "changePassword.messages.validation.newPassword",
+      ),
+      confirmPasswordRequiredMessage: t(
+        "changePassword.messages.validation.confirmPassword",
+      ),
+      hasConfirmPasswordFeature,
+    }),
+  });
+
+  if (hasConfirmPasswordFeature) {
+    ChangePasswordFormSchema = ChangePasswordFormSchema.refine(
       (data) => {
         return data.password === data.confirmPassword;
       },
@@ -44,7 +49,8 @@ export const ChangePasswordForm = ({ handleSubmit, loading }: Properties) => {
         message: t("changePassword.messages.validation.mustMatch"),
         path: ["confirmPassword"],
       },
-    );
+    ) as unknown as typeof ChangePasswordFormSchema;
+  }
 
   return (
     <Provider
@@ -57,7 +63,10 @@ export const ChangePasswordForm = ({ handleSubmit, loading }: Properties) => {
       }
       validationTriggerKey={i18n.language}
     >
-      <ChangePasswordFormFields loading={loading} />
+      <ChangePasswordFormFields
+        hasConfirmPasswordFeature={hasConfirmPasswordFeature}
+        loading={loading}
+      />
     </Provider>
   );
 };