+++ packages/user/src/types/config.ts
@@ -16,6 +16,7 @@ export interface UserConfig {
     [K in keyof (ProtectedRouteOverwrites & PublicRouteOverwrites)]?: string;
   };
   features?: {
+    confirmPassword?: boolean;
     emailVerification?: boolean;
     forgotPassword?: boolean;
     signup?: boolean;
@@ -30,6 +31,7 @@ export interface UserConfig {
     forgotPasswordResendTimeInSeconds?: number;
   };
   homeRoute?: HomeRoute;
+  logoutRedirectRoute?: string;
   socialLoginProviders?: SocialLoginType[];
   supertokens: {
     appName: string;